Use EventDispatcher to allow additional setup of auth backends - move federation auth to federation app
